% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{countrySynonyms}
\alias{countrySynonyms}
\title{Synonyms of country names for each ISO 3 letter country code to enable
conversion.}
\format{A data frame with 281 observations on the following 10 variables.
\describe{ \item{ID}{a numeric vector} \item{ISO3}{ISO 3
letter country code} \item{name1}{country name - most common}
\item{name2}{country name - alternative}
\item{name3}{country name - alternative}
\item{name4}{country name - alternative}
\item{name5}{country name - alternative}
\item{name6}{country name - alternative}
\item{name7}{country name - alternative}
\item{name8}{country name - alternative} }}
\source{
This was derived and used with permission from the Perl Locale
package. \cr Locale::Codes::Country_Codes.\cr Thanks to Sullivan Beck for
pulling this together.\cr Data sources are acknowledged here :\cr
http://search.cpan.org/~sbeck/Locale-Codes-3.23/lib/Locale/Codes/Country.pod
}
\description{
contains a variable number of synonyms (mostly English language) for each
country
}
\details{
This is used by joinCountryData2Map() when country names are used as the
joinCode. Note that using ISO codes is preferable if they are available.
}
\examples{

data(countrySynonyms)


}
\keyword{datasets}

