% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMap.r
\name{getMap}
\alias{getMap}
\title{A simple way to access maps stored in the package.}
\usage{
getMap(resolution = "coarse", projection = NA)
}
\arguments{
\item{resolution}{options "coarse","low","less islands","li","high". For
"high" you need to install the package rworldxtra}

\item{projection}{DEPRECATED OCTOBER 2012 to reproject maps see spTransform
in rgdal}
}
\value{
A SpatialPolygonsDataFrame object.
}
\description{
A simple way to access maps stored in the package.
}
\examples{

plot(getMap())

}
\author{
Barry Rowlingson & Andy South
}
\keyword{misc}

