% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBubbles.r
\name{mapBubbles}
\alias{mapBubbles}
\title{function to produce bubble plots on a map, size and colour determined by
attribute data}
\usage{
mapBubbles(dF = "", nameX = "longitude", nameY = "latitude",
  nameZSize = "", nameZColour = "", fill = TRUE, pch = 21,
  symbolSize = 1, maxZVal = NA, main = nameZSize, numCats = 5,
  catMethod = "categorical", colourPalette = "heat", xlim = NA,
  ylim = NA, mapRegion = "world", borderCol = "grey", oceanCol = NA,
  landCol = NA, addLegend = TRUE, legendBg = "white", legendVals = "",
  legendPos = "bottomright", legendHoriz = FALSE, legendTitle = nameZSize,
  addColourLegend = TRUE, colourLegendPos = "bottomleft",
  colourLegendTitle = nameZColour, add = FALSE, plotZeroVals = TRUE,
  lwd = 0.5, lwdSymbols = 1, ...)
}
\arguments{
\item{dF}{data frame or SpatialPolygonsDataFrame}

\item{nameX}{name of column containing the X variable (longitude), not
needed if dF is a SpatialPolygonsDataFrame}

\item{nameY}{name of column containing the Y variable (lattitude), not
needed if dF is a SpatialPolygonsDataFrame}

\item{nameZSize}{name of column containing numeric variable to set symbol
size}

\item{nameZColour}{name of column containing variable to set symbol colour}

\item{fill}{whether or not to fill symbols TRUE/FALSE}

\item{pch}{symbol type, default of 21 for circles, will work with other
filled symbol types e.g. 22=square, 23=diamond, 24=triangle}

\item{symbolSize}{multiplier of default symbol size}

\item{maxZVal}{the attribute value corresponding to the maximum symbol size,
this can be used to set the scaling the same between multiple plots}

\item{main}{title for the map, set to nameZSize by default}

\item{numCats}{number of categories to put the data in, may be modified if
this number is incompatible with the catMethod chosen}

\item{catMethod}{method for categorisation of data "pretty", "fixedWidth",
"diverging", "logFixedWidth", "quantiles", "categorical", or a numeric
vector defining breaks}

\item{colourPalette}{a string describing the colour palette to use, choice
of : \enumerate{ \item ="palette" for the current palette \item a vector of
valid colours, e.g. =c('red','white','blue') or output from RColourBrewer
\item = one of "heat", "diverging", "white2Black", "black2White", "topo",
"rainbow", "terrain", "negpos8", "negpos9" }}

\item{xlim}{map extents c(west,east), can be overidden by mapRegion}

\item{ylim}{map extents c(south,north), can be overidden by mapRegion}

\item{mapRegion}{a country name from getMap()\$NAME or
'world','africa','oceania','eurasia','uk' sets map extents, overrides
xlim,ylim}

\item{borderCol}{the colour for country borders}

\item{oceanCol}{a colour for the ocean}

\item{landCol}{a colour to fill countries}

\item{addLegend}{whether to add a legend for symbol sizes}

\item{legendBg}{background colour for the legend, NA=transparent}

\item{legendVals}{allows user to set values & hence symbol sizing in legend}

\item{legendPos}{positioning of legend e.g. 'bottomleft', 'topright'}

\item{legendHoriz}{whether to arrange legend elements horizontally
TRUE/FALSE}

\item{legendTitle}{title for the symbol size legend}

\item{addColourLegend}{whether to add a legend for symbol colour}

\item{colourLegendPos}{positioning of colour legend e.g. 'bottomleft',
'topright'}

\item{colourLegendTitle}{title for the colour size legend}

\item{add}{whether to add the symbols to an existing map, TRUE/FALSE}

\item{plotZeroVals}{whether to plot zero values as a cross, TRUE/FALSE}

\item{lwd}{line width for country borders}

\item{lwdSymbols}{line width for symbols}

\item{\dots}{any extra arguments to points()}
}
\value{
currently doesn't return anything
}
\description{
The function will produce a map with bubbles (circles) centred on country
centroids (or other chosen points). Bubbles can be sized and coloured
according to specified attribute values.
}
\details{
By default separate legends are added fro symbol size and colouring on
either side of the plot, these can be modified by altering legend
parameters.
}
\examples{

mapBubbles()
#square symbols
mapBubbles(pch=22)

mapBubbles(dF=getMap(), nameZSize="POP_EST", nameZColour="GEO3")

#change colour
mapBubbles(dF=getMap(), nameZSize="POP_EST", nameZColour="GEO3"
          ,colourPalette='rainbow', oceanCol='lightblue', landCol='wheat') 


data("countryExData",envir=environment(),package="rworldmap")
sPDF <- joinCountryData2Map(countryExData,joinCode = "ISO3"
                           ,nameJoinColumn = "ISO3V10")
                           
mapBubbles(sPDF, nameZSize="POP_EST",nameZColour="BIODIVERSITY"
          ,colourPalette='topo',numCats=5,catMethod="quantiles")

#filled bubbles with set transparency
mapBubbles(fill=TRUE,colourPalette=adjustcolor(palette(), alpha.f = 0.5))
#add bubble edge of a single colour (also with option to set transparency
mapBubbles(nameZColour = adjustcolor('black', alpha.f = 0.7), fill=FALSE, add=TRUE)

}
\author{
andy south
}
\keyword{aplot}

