% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwmCheckAndLoadInput.R
\name{rwmCheckAndLoadInput}
\alias{rwmCheckAndLoadInput}
\title{internal function to check and load input data to mapping functions}
\usage{
rwmCheckAndLoadInput(inputData = "", inputNeeded = "sPDF",
  callingFunction = "")
}
\arguments{
\item{inputData}{a dF, sPDF or "", for latter an internal example data is
used}

\item{inputNeeded}{"sPDF", "sPDF or dF", "dF"}

\item{callingFunction}{optional : name of the calling function}
}
\value{
invisibly returns a dF or sPDF
}
\description{
Internal function checking and loading dFs or sPDFs to
\code{\link{mapCountryData}}, \code{\link{mapPolys}}, \code{\link{mapPies}},
\code{\link{mapBubbles}}, \code{\link{mapBars}}.
}
\details{
a rworldmap internal function, unlikely to be of use to users
}
\author{
andy south
}
\keyword{aplot}

