% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rworldmapExamples.R
\name{rworldmapExamples}
\alias{rworldmapExamples}
\title{Example code for plot creation}
\usage{
rworldmapExamples()
}
\description{
Example code to demonstrate creation of a series of plots
}
\author{
andy south
}
\keyword{aplot}

